requirements.txt
asgiref==3.3.4
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
crontab==0.23.0
Django==3.0.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.4
dnspython==2.2.1
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
numpy==1.23.5
outcome==1.2.0
packaging==22.0
pandas==1.5.2
passlib==1.7.4
PyJWT==1.7.1
pymongo==3.11.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2021.1
requests==2.28.1
selenium==4.7.2
sendgrid==6.9.7
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
starkbank-ecdsa==2.2.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
undetected-chromedriver==3.1.7
urllib3==1.26.13
webdriver-manager==3.8.5
websockets==10.4
wsproto==1.2.0
